moves around files for clarity and readability
