+++ client/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import store from './utils/storeProvider';
 
-import Index from './pages/index';
+import Index from './routes/index';
 import './sass/base.scss';
 
 const rootElement = document.getElementById('root');
